Run GWSkyNet tasks on openmp queue
Originally, we allocated a dedicated queue for GWSkyNet tasks
because they needed a lot of memory and because they used many
CPUs and the flavor of parallelism was poorly understood (#514).

Their memory usage is now greatly reduced by using multi-resolution
sky maps and the flavor of parallelism is now well understood as
resulting from OpenMP usage in ligo.skymap volume rendering
(manleong.chan/gwskynet#5).

It is now safe to reallocate GWSkyNet tasks to the workers that we
have dedicated for OpenMP parallel workflows. This patch directs
GWSkyNet tasks to the `openmp` queue and retires the now unnecssary
`skynet` queue. This simplifies deployment while also reducing
resource requirements, both on the cluster and in Kubernetes
development setups.

There is a FIXME comment noting that even the high CPU usage
stemming from the OpenMP-parallel volume rendering may eventually
be unnecessary (see manleong.chan/gwskynet#6).